0203/2-3-main (1)/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.128.0",
    "mariadb>=1.1.14",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
]
